American Community Survey population data

These data sets contain population data for Kansas City, Missouri, sourced from the US Census Bureau's American Community Survey (ACS). In 1-year ACS data sets, the summary level is Kansas City only. In 5-year ACS data sets, available summary levels include city, census tract, and ZIP code tabulation area (ZCTA).

Data sets were downloaded from data.census.gov using the tidycensus package. All data variables from the full and collapsed Age and Sex, Race, and Hispanic Origin tables are included (i.e., tables beginning with B01, B02, B03, C01, C02, and C03).

See Guidance for Data Users for information on how to use ACS data.

Usage

acs1_2014

acs1_2015

acs1_2016

acs1_2017

acs1_2018

acs1_2019

acs1_2021

acs1_2022

acs1_2023

acs5_city_2014

acs5_city_2015

acs5_city_2016

acs5_city_2017

acs5_city_2018

acs5_city_2019

acs5_city_2020

acs5_city_2021

acs5_city_2022

acs5_city_2023

acs5_tract_2014

acs5_tract_2015

acs5_tract_2016

acs5_tract_2017

acs5_tract_2018

acs5_tract_2019

acs5_tract_2020

acs5_tract_2021

acs5_tract_2022

acs5_tract_2023

acs5_zcta_2014

acs5_zcta_2015

acs5_zcta_2016

acs5_zcta_2017

acs5_zcta_2018

acs5_zcta_2019

acs5_zcta_2020

acs5_zcta_2021

acs5_zcta_2022

acs5_zcta_2023

Format

GEOID

Geographic identifier

Data type: character

NAME

Geographic area name

Data type: character

variable

Variable code (ACS table and variable codes separated by an underscore)

Data type: character

estimate

Estimated population count

Data type: double

moe

Margin of error

Data type: double

label

Descriptive variable label

Data type: character

concept

ACS table name

Data type: character

geography

Smallest geography for which a variable is summarized

Data type: character

5-year ACS tables only

An object of class tbl_df (inherits from tbl, data.frame) with 826 rows and 7 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 826 rows and 7 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 826 rows and 7 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 826 rows and 7 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 826 rows and 7 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 826 rows and 7 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 850 rows and 7 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 821 rows and 7 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 529 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 630 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 630 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 630 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 630 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 630 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 630 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 630 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 667 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 647 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 107916 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 128520 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 128520 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 128520 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 128520 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 128520 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 140490 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 140490 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 148741 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 144281 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 36501 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 43470 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 43470 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 43470 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 43470 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 43470 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 48510 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 48510 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 51359 rows and 8 columns.

An object of class tbl_df (inherits from tbl, data.frame) with 49819 rows and 8 columns.

Source

US Census Bureau American Community Survey 1-Year and 5-Year data sets (2014-2023)

Examples

acs1_2023
#> # A tibble: 821 × 7
#>    GEOID   NAME                       variable   estimate   moe label    concept
#>    <chr>   <chr>                      <chr>         <dbl> <dbl> <chr>    <chr>  
#>  1 2938000 Kansas City city, Missouri B01001_001   510671    69 Estimat… Sex by…
#>  2 2938000 Kansas City city, Missouri B01001_002   250013  4299 Estimat… Sex by…
#>  3 2938000 Kansas City city, Missouri B01001_003    16073  1804 Estimat… Sex by…
#>  4 2938000 Kansas City city, Missouri B01001_004    16430  2387 Estimat… Sex by…
#>  5 2938000 Kansas City city, Missouri B01001_005    14115  2267 Estimat… Sex by…
#>  6 2938000 Kansas City city, Missouri B01001_006     9759  1566 Estimat… Sex by…
#>  7 2938000 Kansas City city, Missouri B01001_007     5856  1104 Estimat… Sex by…
#>  8 2938000 Kansas City city, Missouri B01001_008     1616   709 Estimat… Sex by…
#>  9 2938000 Kansas City city, Missouri B01001_009     2999  1029 Estimat… Sex by…
#> 10 2938000 Kansas City city, Missouri B01001_010    11142  2237 Estimat… Sex by…
#> # ℹ 811 more rows

acs5_tract_2023
#> # A tibble: 144,281 × 8
#>    GEOID       NAME              variable estimate   moe label concept geography
#>    <chr>       <chr>             <chr>       <dbl> <dbl> <chr> <chr>   <chr>    
#>  1 29037060001 Census Tract 600… B01001_…     4927   542 Esti… Sex by… block gr…
#>  2 29037060001 Census Tract 600… B01001_…     2518   335 Esti… Sex by… block gr…
#>  3 29037060001 Census Tract 600… B01001_…       90    59 Esti… Sex by… block gr…
#>  4 29037060001 Census Tract 600… B01001_…      130    73 Esti… Sex by… block gr…
#>  5 29037060001 Census Tract 600… B01001_…      180   117 Esti… Sex by… block gr…
#>  6 29037060001 Census Tract 600… B01001_…      154    70 Esti… Sex by… block gr…
#>  7 29037060001 Census Tract 600… B01001_…       23    23 Esti… Sex by… block gr…
#>  8 29037060001 Census Tract 600… B01001_…       21    32 Esti… Sex by… block gr…
#>  9 29037060001 Census Tract 600… B01001_…        7    10 Esti… Sex by… block gr…
#> 10 29037060001 Census Tract 600… B01001_…      121    83 Esti… Sex by… block gr…
#> # ℹ 144,271 more rows