Decennial census population data

These data sets contain population data for Kansas City, Missouri, sourced from the US Census Bureau's decennial census. Summary levels include city, county, census tract, census block group, census block, and ZIP code tabulation area (ZCTA).

Data sets were downloaded from data.census.gov using get_kc_pop(). See Guidance for Decennial Census of Population and Housing Data Users for information on how to use decennial census data.

Usage

dec_city_2010

dec_city_2020

dec_county_2010

dec_county_2020

dec_tract_2010

dec_tract_2020

dec_blockgroup_2010

dec_blockgroup_2020

dec_block_2010

dec_block_2020

dec_zcta_2010

dec_zcta_2020

Format

GEOID

Geographic identifier

Data type: character

NAME

Geographic area name

Data type: character

variable

Variable code

Data type: character

value

Population count

Data type: double

label

Descriptive variable label

Data type: character

concept

Census table name

Data type: character

An object of class data.frame with 1 rows and 6 columns.

An object of class data.frame with 4 rows and 6 columns.

An object of class data.frame with 4 rows and 6 columns.

An object of class data.frame with 173 rows and 6 columns.

An object of class data.frame with 185 rows and 6 columns.

An object of class data.frame with 460 rows and 6 columns.

An object of class data.frame with 510 rows and 6 columns.

An object of class data.frame with 12001 rows and 6 columns.

An object of class data.frame with 10730 rows and 6 columns.

An object of class data.frame with 62 rows and 6 columns.

An object of class data.frame with 59 rows and 6 columns.

Source

US Census Bureau decennial census

Examples

dec_city_2010
#>     GEOID                       NAME variable  value label    concept
#> 1 2938000 Kansas City city, Missouri  P012001 459787 Total SEX BY AGE

dec_tract_2020
#>           GEOID                                           NAME variable value
#> 1   29095001800      Census Tract 18; Jackson County; Missouri P12_001N  3161
#> 2   29095000900       Census Tract 9; Jackson County; Missouri P12_001N  3005
#> 3   29095001000      Census Tract 10; Jackson County; Missouri P12_001N  2910
#> 4   29095001100      Census Tract 11; Jackson County; Missouri P12_001N  2378
#> 5   29095001900      Census Tract 19; Jackson County; Missouri P12_001N  3380
#> 6   29095002000      Census Tract 20; Jackson County; Missouri P12_001N  2315
#> 7   29095002100      Census Tract 21; Jackson County; Missouri P12_001N  4470
#> 8   29095002200      Census Tract 22; Jackson County; Missouri P12_001N  3224
#> 9   29095002300      Census Tract 23; Jackson County; Missouri P12_001N  2149
#> 10  29095003400      Census Tract 34; Jackson County; Missouri P12_001N  3288
#> 11  29095003700      Census Tract 37; Jackson County; Missouri P12_001N  1049
#> 12  29095003800      Census Tract 38; Jackson County; Missouri P12_001N  1538
#> 13  29095004300      Census Tract 43; Jackson County; Missouri P12_001N  2794
#> 14  29095004400      Census Tract 44; Jackson County; Missouri P12_001N  1495
#> 15  29095004600      Census Tract 46; Jackson County; Missouri P12_001N  2689
#> 16  29095005100      Census Tract 51; Jackson County; Missouri P12_001N  1620
#> 17  29095005200      Census Tract 52; Jackson County; Missouri P12_001N  1038
#> 18  29095005300      Census Tract 53; Jackson County; Missouri P12_001N  1437
#> 19  29095005400      Census Tract 54; Jackson County; Missouri P12_001N   826
#> 20  29095005500      Census Tract 55; Jackson County; Missouri P12_001N   927
#> 21  29095005601   Census Tract 56.01; Jackson County; Missouri P12_001N  1726
#> 22  29095005602   Census Tract 56.02; Jackson County; Missouri P12_001N  1420
#> 23  29095005700      Census Tract 57; Jackson County; Missouri P12_001N  2285
#> 24  29095005801   Census Tract 58.01; Jackson County; Missouri P12_001N  2172
#> 25  29095006000      Census Tract 60; Jackson County; Missouri P12_001N  1478
#> 26  29095006100      Census Tract 61; Jackson County; Missouri P12_001N  2695
#> 27  29095006300      Census Tract 63; Jackson County; Missouri P12_001N  1737
#> 28  29095006500      Census Tract 65; Jackson County; Missouri P12_001N  1632
#> 29  29095006600      Census Tract 66; Jackson County; Missouri P12_001N  1833
#> 30  29095006700      Census Tract 67; Jackson County; Missouri P12_001N  1807
#> 31  29095006900      Census Tract 69; Jackson County; Missouri P12_001N  1338
#> 32  29095007100      Census Tract 71; Jackson County; Missouri P12_001N  2993
#> 33  29095007200      Census Tract 72; Jackson County; Missouri P12_001N  2044
#> 34  29095007301   Census Tract 73.01; Jackson County; Missouri P12_001N  1818
#> 35  29095007302   Census Tract 73.02; Jackson County; Missouri P12_001N  1377
#> 36  29095007400      Census Tract 74; Jackson County; Missouri P12_001N  4397
#> 37  29095007500      Census Tract 75; Jackson County; Missouri P12_001N  2727
#> 38  29095007600      Census Tract 76; Jackson County; Missouri P12_001N  2294
#> 39  29095007700      Census Tract 77; Jackson County; Missouri P12_001N  1665
#> 40  29095007802   Census Tract 78.02; Jackson County; Missouri P12_001N  2258
#> 41  29095007900      Census Tract 79; Jackson County; Missouri P12_001N  3498
#> 42  29095008000      Census Tract 80; Jackson County; Missouri P12_001N  3022
#> 43  29095008100      Census Tract 81; Jackson County; Missouri P12_001N  2113
#> 44  29095008200      Census Tract 82; Jackson County; Missouri P12_001N  2913
#> 45  29095017100     Census Tract 171; Jackson County; Missouri P12_001N  3814
#> 46  29095016802  Census Tract 168.02; Jackson County; Missouri P12_001N  1338
#> 47  29095016900     Census Tract 169; Jackson County; Missouri P12_001N  2361
#> 48  29095017000     Census Tract 170; Jackson County; Missouri P12_001N  2510
#> 49  29095017200     Census Tract 172; Jackson County; Missouri P12_001N  4573
#> 50  29095017300     Census Tract 173; Jackson County; Missouri P12_001N  3354
#> 51  29095017400     Census Tract 174; Jackson County; Missouri P12_001N  2807
#> 52  29095017500     Census Tract 175; Jackson County; Missouri P12_001N  2012
#> 53  29095017600     Census Tract 176; Jackson County; Missouri P12_001N  5118
#> 54  29095017800     Census Tract 178; Jackson County; Missouri P12_001N  3809
#> 55  29095018102  Census Tract 181.02; Jackson County; Missouri P12_001N  3148
#> 56  29095018200     Census Tract 182; Jackson County; Missouri P12_001N  4166
#> 57  29095980101 Census Tract 9801.01; Jackson County; Missouri P12_001N     3
#> 58  29095980802 Census Tract 9808.02; Jackson County; Missouri P12_001N     1
#> 59  29095988300    Census Tract 9883; Jackson County; Missouri P12_001N    97
#> 60  29047020902     Census Tract 209.02; Clay County; Missouri P12_001N  2562
#> 61  29047020901     Census Tract 209.01; Clay County; Missouri P12_001N  6107
#> 62  29047021204     Census Tract 212.04; Clay County; Missouri P12_001N  5679
#> 63  29047021208     Census Tract 212.08; Clay County; Missouri P12_001N  6058
#> 64  29047021209     Census Tract 212.09; Clay County; Missouri P12_001N  4461
#> 65  29047021210     Census Tract 212.10; Clay County; Missouri P12_001N  4368
#> 66  29047021211     Census Tract 212.11; Clay County; Missouri P12_001N  2804
#> 67  29047021212     Census Tract 212.12; Clay County; Missouri P12_001N  6754
#> 68  29047021213     Census Tract 212.13; Clay County; Missouri P12_001N  3193
#> 69  29047021214     Census Tract 212.14; Clay County; Missouri P12_001N  1417
#> 70  29047021303     Census Tract 213.03; Clay County; Missouri P12_001N  5065
#> 71  29047021307     Census Tract 213.07; Clay County; Missouri P12_001N  7622
#> 72  29047021309     Census Tract 213.09; Clay County; Missouri P12_001N  5911
#> 73  29047021311     Census Tract 213.11; Clay County; Missouri P12_001N  2866
#> 74  29047021312     Census Tract 213.12; Clay County; Missouri P12_001N  7620
#> 75  29047021313     Census Tract 213.13; Clay County; Missouri P12_001N  8275
#> 76  29047021314     Census Tract 213.14; Clay County; Missouri P12_001N  7251
#> 77  29095008600      Census Tract 86; Jackson County; Missouri P12_001N  5385
#> 78  29095008300      Census Tract 83; Jackson County; Missouri P12_001N  2337
#> 79  29095008400      Census Tract 84; Jackson County; Missouri P12_001N  2800
#> 80  29095008500      Census Tract 85; Jackson County; Missouri P12_001N  3573
#> 81  29095008700      Census Tract 87; Jackson County; Missouri P12_001N  2558
#> 82  29095008800      Census Tract 88; Jackson County; Missouri P12_001N  3430
#> 83  29095008900      Census Tract 89; Jackson County; Missouri P12_001N  1950
#> 84  29095009000      Census Tract 90; Jackson County; Missouri P12_001N  4219
#> 85  29095009100      Census Tract 91; Jackson County; Missouri P12_001N  3735
#> 86  29095009200      Census Tract 92; Jackson County; Missouri P12_001N  2890
#> 87  29095009300      Census Tract 93; Jackson County; Missouri P12_001N  2394
#> 88  29095009400      Census Tract 94; Jackson County; Missouri P12_001N  4391
#> 89  29095009500      Census Tract 95; Jackson County; Missouri P12_001N  3160
#> 90  29095009600      Census Tract 96; Jackson County; Missouri P12_001N  1555
#> 91  29095009700      Census Tract 97; Jackson County; Missouri P12_001N  1592
#> 92  29095009800      Census Tract 98; Jackson County; Missouri P12_001N  3484
#> 93  29095009900      Census Tract 99; Jackson County; Missouri P12_001N  2130
#> 94  29095010001  Census Tract 100.01; Jackson County; Missouri P12_001N  1801
#> 95  29095010002  Census Tract 100.02; Jackson County; Missouri P12_001N  3575
#> 96  29095010103  Census Tract 101.03; Jackson County; Missouri P12_001N  2417
#> 97  29095010105  Census Tract 101.05; Jackson County; Missouri P12_001N  3510
#> 98  29095010201  Census Tract 102.01; Jackson County; Missouri P12_001N  1836
#> 99  29095010203  Census Tract 102.03; Jackson County; Missouri P12_001N  4042
#> 100 29095010204  Census Tract 102.04; Jackson County; Missouri P12_001N  4379
#> 101 29095010500     Census Tract 105; Jackson County; Missouri P12_001N  4803
#> 102 29095010600     Census Tract 106; Jackson County; Missouri P12_001N  3044
#> 103 29095010702  Census Tract 107.02; Jackson County; Missouri P12_001N  3811
#> 104 29095012802  Census Tract 128.02; Jackson County; Missouri P12_001N  3772
#> 105 29095012903  Census Tract 129.03; Jackson County; Missouri P12_001N  3577
#> 106 29095012904  Census Tract 129.04; Jackson County; Missouri P12_001N  2645
#> 107 29095012906  Census Tract 129.06; Jackson County; Missouri P12_001N  4562
#> 108 29095013003  Census Tract 130.03; Jackson County; Missouri P12_001N  4842
#> 109 29095013100     Census Tract 131; Jackson County; Missouri P12_001N  2983
#> 110 29095013203  Census Tract 132.03; Jackson County; Missouri P12_001N  4102
#> 111 29095013208  Census Tract 132.08; Jackson County; Missouri P12_001N  3022
#> 112 29095013210  Census Tract 132.10; Jackson County; Missouri P12_001N  2564
#> 113 29095013405  Census Tract 134.05; Jackson County; Missouri P12_001N  1981
#> 114 29095013407  Census Tract 134.07; Jackson County; Missouri P12_001N  2574
#> 115 29095013417  Census Tract 134.17; Jackson County; Missouri P12_001N  2918
#> 116 29095013418  Census Tract 134.18; Jackson County; Missouri P12_001N  2742
#> 117 29095013504  Census Tract 135.04; Jackson County; Missouri P12_001N  3488
#> 118 29037060100        Census Tract 601; Cass County; Missouri P12_001N  1413
#> 119 29165030002   Census Tract 300.02; Platte County; Missouri P12_001N  3392
#> 120 29165030003   Census Tract 300.03; Platte County; Missouri P12_001N  2554
#> 121 29047020201     Census Tract 202.01; Clay County; Missouri P12_001N  5423
#> 122 29047020202     Census Tract 202.02; Clay County; Missouri P12_001N  4867
#> 123 29047020300        Census Tract 203; Clay County; Missouri P12_001N  6688
#> 124 29047020400        Census Tract 204; Clay County; Missouri P12_001N  2561
#> 125 29047020500        Census Tract 205; Clay County; Missouri P12_001N  6252
#> 126 29047020602     Census Tract 206.02; Clay County; Missouri P12_001N  4417
#> 127 29047020603     Census Tract 206.03; Clay County; Missouri P12_001N  4196
#> 128 29047020604     Census Tract 206.04; Clay County; Missouri P12_001N  5484
#> 129 29165030004   Census Tract 300.04; Platte County; Missouri P12_001N  5920
#> 130 29165030101   Census Tract 301.01; Platte County; Missouri P12_001N  3831
#> 131 29165030102   Census Tract 301.02; Platte County; Missouri P12_001N  4775
#> 132 29165030103   Census Tract 301.03; Platte County; Missouri P12_001N  5850
#> 133 29165030201   Census Tract 302.01; Platte County; Missouri P12_001N  5492
#> 134 29165030207   Census Tract 302.07; Platte County; Missouri P12_001N  6102
#> 135 29165030208   Census Tract 302.08; Platte County; Missouri P12_001N  3165
#> 136 29165030210   Census Tract 302.10; Platte County; Missouri P12_001N  2067
#> 137 29165030211   Census Tract 302.11; Platte County; Missouri P12_001N  3011
#> 138 29165030212   Census Tract 302.12; Platte County; Missouri P12_001N  4849
#> 139 29165030213   Census Tract 302.13; Platte County; Missouri P12_001N  2722
#> 140 29165030214   Census Tract 302.14; Platte County; Missouri P12_001N  2477
#> 141 29165030215   Census Tract 302.15; Platte County; Missouri P12_001N  1552
#> 142 29165030216   Census Tract 302.16; Platte County; Missouri P12_001N  4194
#> 143 29165030306   Census Tract 303.06; Platte County; Missouri P12_001N  5074
#> 144 29165030308   Census Tract 303.08; Platte County; Missouri P12_001N  5728
#> 145 29165030310   Census Tract 303.10; Platte County; Missouri P12_001N  3755
#> 146 29165030601   Census Tract 306.01; Platte County; Missouri P12_001N  1909
#> 147 29165030602   Census Tract 306.02; Platte County; Missouri P12_001N  5886
#> 148 29165980000     Census Tract 9800; Platte County; Missouri P12_001N    12
#> 149 29095014300     Census Tract 143; Jackson County; Missouri P12_001N  5483
#> 150 29095014205  Census Tract 142.05; Jackson County; Missouri P12_001N  3725
#> 151 29095014206  Census Tract 142.06; Jackson County; Missouri P12_001N  5093
#> 152 29095014400     Census Tract 144; Jackson County; Missouri P12_001N  2951
#> 153 29095014501  Census Tract 145.01; Jackson County; Missouri P12_001N  4106
#> 154 29095014504  Census Tract 145.04; Jackson County; Missouri P12_001N  3538
#> 155 29095015200     Census Tract 152; Jackson County; Missouri P12_001N  2646
#> 156 29095015300     Census Tract 153; Jackson County; Missouri P12_001N  3010
#> 157 29095015401  Census Tract 154.01; Jackson County; Missouri P12_001N  2138
#> 158 29095015402  Census Tract 154.02; Jackson County; Missouri P12_001N  1341
#> 159 29095015500     Census Tract 155; Jackson County; Missouri P12_001N  1328
#> 160 29095015600     Census Tract 156; Jackson County; Missouri P12_001N  2483
#> 161 29095015701  Census Tract 157.01; Jackson County; Missouri P12_001N  2219
#> 162 29095015702  Census Tract 157.02; Jackson County; Missouri P12_001N  3191
#> 163 29095015800     Census Tract 158; Jackson County; Missouri P12_001N  1932
#> 164 29095015900     Census Tract 159; Jackson County; Missouri P12_001N  1955
#> 165 29095016000     Census Tract 160; Jackson County; Missouri P12_001N  1756
#> 166 29095016100     Census Tract 161; Jackson County; Missouri P12_001N  1906
#> 167 29095016200     Census Tract 162; Jackson County; Missouri P12_001N  2052
#> 168 29095016300     Census Tract 163; Jackson County; Missouri P12_001N  2082
#> 169 29095016400     Census Tract 164; Jackson County; Missouri P12_001N  1741
#> 170 29095016500     Census Tract 165; Jackson County; Missouri P12_001N  1868
#> 171 29095016600     Census Tract 166; Jackson County; Missouri P12_001N  1619
#> 172 29095016700     Census Tract 167; Jackson County; Missouri P12_001N  3366
#> 173 29095016801  Census Tract 168.01; Jackson County; Missouri P12_001N  2178
#> 174 29095000300       Census Tract 3; Jackson County; Missouri P12_001N  2087
#> 175 29095000600       Census Tract 6; Jackson County; Missouri P12_001N  4653
#> 176 29095000700       Census Tract 7; Jackson County; Missouri P12_001N  4056
#> 177 29095000800       Census Tract 8; Jackson County; Missouri P12_001N  4334
#> 178 29047021808     Census Tract 218.08; Clay County; Missouri P12_001N  4648
#> 179 29047021807     Census Tract 218.07; Clay County; Missouri P12_001N  4233
#> 180 29047021810     Census Tract 218.10; Clay County; Missouri P12_001N  3033
#> 181 29047021900        Census Tract 219; Clay County; Missouri P12_001N  6768
#> 182 29047022100        Census Tract 221; Clay County; Missouri P12_001N  5343
#> 183 29047022200        Census Tract 222; Clay County; Missouri P12_001N  4110
#> 184 29095013612  Census Tract 136.12; Jackson County; Missouri P12_001N  7107
#> 185 29095014203  Census Tract 142.03; Jackson County; Missouri P12_001N  4155
#>         label                                concept
#> 1    !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 2    !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 3    !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 4    !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 5    !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 6    !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 7    !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 8    !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 9    !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 10   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 11   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 12   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 13   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 14   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 15   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 16   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 17   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 18   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 19   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 20   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 21   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 22   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 23   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 24   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 25   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 26   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 27   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 28   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 29   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 30   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 31   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 32   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 33   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 34   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 35   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 36   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 37   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 38   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 39   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 40   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 41   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 42   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 43   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 44   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 45   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 46   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 47   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 48   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 49   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 50   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 51   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 52   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 53   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 54   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 55   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 56   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 57   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 58   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 59   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 60   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 61   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 62   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 63   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 64   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 65   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 66   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 67   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 68   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 69   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 70   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 71   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 72   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 73   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 74   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 75   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 76   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 77   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 78   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 79   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 80   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 81   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 82   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 83   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 84   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 85   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 86   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 87   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 88   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 89   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 90   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 91   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 92   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 93   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 94   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 95   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 96   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 97   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 98   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 99   !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 100  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 101  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 102  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 103  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 104  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 105  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 106  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 107  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 108  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 109  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 110  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 111  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 112  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 113  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 114  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 115  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 116  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 117  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 118  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 119  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 120  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 121  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 122  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 123  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 124  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 125  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 126  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 127  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 128  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 129  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 130  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 131  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 132  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 133  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 134  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 135  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 136  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 137  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 138  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 139  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 140  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 141  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 142  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 143  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 144  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 145  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 146  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 147  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 148  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 149  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 150  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 151  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 152  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 153  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 154  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 155  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 156  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 157  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 158  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 159  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 160  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 161  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 162  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 163  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 164  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 165  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 166  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 167  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 168  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 169  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 170  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 171  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 172  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 173  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 174  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 175  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 176  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 177  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 178  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 179  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 180  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 181  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 182  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 183  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 184  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES
#> 185  !!Total: SEX BY AGE FOR SELECTED AGE CATEGORIES