Census block group shapefiles

TIGER/Line shapefiles for census block groups in Kansas City, Missouri, downloaded from census.gov and modified using get_kc_sf().

The US Census Bureau uses the Global Coordinate System North American Datum of 1983 (GCS NAD83/EPSG: 4269).

Usage

sf_blockgroup_2018

sf_blockgroup_2019

sf_blockgroup_2020

sf_blockgroup_2021

sf_blockgroup_2022

sf_blockgroup_2023

sf_blockgroup_2024

Format

TBD

An object of class sf (inherits from data.frame) with 451 rows and 16 columns.

An object of class sf (inherits from data.frame) with 510 rows and 16 columns.

An object of class sf (inherits from data.frame) with 510 rows and 16 columns.

An object of class sf (inherits from data.frame) with 510 rows and 16 columns.

An object of class sf (inherits from data.frame) with 510 rows and 17 columns.

An object of class sf (inherits from data.frame) with 510 rows and 17 columns.

Source

US Census Bureau TIGER/Line Shapefiles (2018-2024)