City boundary shapefiles

TIGER/Line shapefiles for the city of Kansas City, Missouri, downloaded from census.gov using the tigris package.

The US Census Bureau uses the Global Coordinate System North American Datum of 1983 (GCS NAD83/EPSG: 4269).

Usage

shp_city_2014

shp_city_2015

shp_city_2016

shp_city_2017

shp_city_2018

shp_city_2019

shp_city_2020

shp_city_2021

shp_city_2022

shp_city_2023

Format

STATEFP

State FIPS code

Data type: character

PLACEFP

Place FIPS code

Data type: character

PLACENS

Place NS code

Data type: character

GEOID

Geographic identifier

Data type: character

GEOIDFQ

Fully qualified GEOID

Data type: character

2023 table only

NAME

Geographic area name

Data type: character

NAMELSAD

Geographic area name and legal/statistical area description

Data type: character

LSAD

Legal/statistical area description

Data type: character

CLASSFP

FIPS class code

Data type: character

PCICBSA

Core based statistical area principal city indicator

Data type: character

PCINECTA

New England city and town area principal city indicator

Data type: character

2014-2022 tables only

MTFCC

MAF/TIGER Feature Class Code

Data type: character

FUNCSTAT

Legal functional status

Data type: character

ALAND

Land area

Data type: numeric

AWATER

Water area

Data type: numeric

INTPTLAT

Latitude of the internal point

Data type: character

INTPTLON

Longitude of the internal point

Data type: character

geometry

TBD

Data type: character

An object of class sf (inherits from data.frame) with 1 rows and 17 columns.

An object of class sf (inherits from data.frame) with 1 rows and 17 columns.

An object of class sf (inherits from data.frame) with 1 rows and 17 columns.

An object of class sf (inherits from data.frame) with 1 rows and 17 columns.

An object of class sf (inherits from data.frame) with 1 rows and 17 columns.

An object of class sf (inherits from data.frame) with 1 rows and 17 columns.

An object of class sf (inherits from data.frame) with 1 rows and 17 columns.

An object of class sf (inherits from data.frame) with 1 rows and 17 columns.

An object of class sf (inherits from data.frame) with 1 rows and 17 columns.

Source

US Census Bureau TIGER/Line Shapefiles (2014-2023)